backend/requirements.txt
# FastAPI and core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# MongoDB (Motor async driver)
motor>=3.3.0
pymongo>=4.5.0
bson>=0.5.10

# Testing with MongoDB mocks
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
mongomock>=4.1.2
mongomock-motor>=0.0.21

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
aiohttp>=3.8.0

# Web3 and blockchain
web3>=6.15.1
eth-account>=0.9.0

# Cryptography
cryptography>=41.0.0

# Development
black>=23.0.0
flake8>=6.0.0
